fixed issue with using CTime::GetCurrent time as an int64 rather than an int as the string.Format was expecting, caused duplicate clips that were found to not be moved to the top of the list
git-svn-id: svn://svn.code.sf.net/p/ditto-cp/code/trunk@642 595ec19a-5cb4-439b-94a8-42fb3063c22c

+++ CopyProperties.cpp
@@ -304,7 +304,7 @@ void CCopyProperties::LoadDataIntoCClip(CClip &Clip)
 	//don't auto delete check box is checked
 	if(m_bNeverAutoDelete)
 	{
-		Clip.m_dontAutoDelete = (long)CTime::GetCurrentTime().GetTime();
+		Clip.m_dontAutoDelete = (int)CTime::GetCurrentTime().GetTime();
 	}
 	else if(m_bNeverAutoDelete == FALSE)
 	{
